% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_restaurants.R
\name{compare_restaurants}
\alias{compare_restaurants}
\title{Compare restaurants by average calories}
\usage{
compare_restaurants(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A table showing average calories by restaurant
}
\description{
This function calculates the average calories for each restaurant.
}
