% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood.R
\docType{data}
\name{fastfood}
\alias{fastfood}
\title{Fast Food Nutrition Dataset}
\format{
A data frame with 9 rows and 7 variables:
\describe{
\item{restaurant}{Restaurant name}
\item{item}{Menu item name}
\item{calories}{Calories}
\item{protein}{Protein}
\item{sugar}{Sugar}
\item{fat}{Fat}
\item{sodium}{Sodium}
}
}
\usage{
fastfood
}
\description{
A dataset containing nutrition information for fast food menu items.
}
\keyword{datasets}
