% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_calories.R
\name{highest_calories}
\alias{highest_calories}
\title{Find the highest calorie item}
\usage{
highest_calories(data)
}
\arguments{
\item{data}{A dataframe with fast food nutrition info}
}
\value{
A row showing the item with the highest calories
}
\description{
This function returns the menu item with the most calories.
}
