% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_protein.R
\name{highest_protein}
\alias{highest_protein}
\title{Find the highest protein item}
\usage{
highest_protein(data)
}
\arguments{
\item{data}{A dataframe with nutrition info}
}
\value{
A row showing the item with the highest protein
}
\description{
This function returns the menu item with the most protein.
}
