% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_sodium.R
\name{highest_sodium}
\alias{highest_sodium}
\title{Find the highest sodium item}
\usage{
highest_sodium(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A row with the highest sodium value
}
\description{
This function returns the item with the most sodium.
}
