% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowest_calories.R
\name{lowest_calories}
\alias{lowest_calories}
\title{Find the lowest calorie item}
\usage{
lowest_calories(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A row with the lowest calorie item
}
\description{
This function returns the item with the least calories.
}
