% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowest_sugar.R
\name{lowest_sugar}
\alias{lowest_sugar}
\title{Find the lowest sugar item}
\usage{
lowest_sugar(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A row with the lowest sugar item
}
\description{
This function returns the item with the least sugar.
}
