% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calories.R
\name{plot_calories}
\alias{plot_calories}
\title{Plot calories for each item}
\usage{
plot_calories(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
This function creates a simple bar chart of calories
for each menu item.
}
