% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant_summary.R
\name{restaurant_summary}
\alias{restaurant_summary}
\title{Get a summary of each restaurant}
\usage{
restaurant_summary(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A summary table by restaurant
}
\description{
This function gives the average calories, protein, sugar,
fat, and sodium for each restaurant.
}
